backend/contributors/danrodjim/people/requirements.txt
Flask
marshmallow_sqlalchemy
pandas